requirements.txt
aiogram==2.7
asyncpg==0.29.0
SQLAlchemy==1.4.49
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.29.0
